Damage uses doubles now, CharacterSpecies has default stat modifiers.

+++ characters/nodes/common_components/survival/armor_component.h
@@ -12,8 +12,8 @@ protected:
 	static void _bind_methods();
 	void _notification(int p_what);
 
-	int64_t armor = 25;
-	int64_t max_armor = 25;
+	double armor = 25;
+	double max_armor = 25;
 
 	double gen_hit_delay = 10.0;
 	double gen_tick_rate = 1.0; // if 0, always be max after hit delay.
@@ -30,17 +30,17 @@ public:
 
 	void handle_damage(Ref<DamageSource> source);
 
-	void set_armor(int64_t value);
-	int64_t get_armor();
+	void set_armor(double value);
+	double get_armor();
 
-	void set_max_armor(int64_t value);
-	int64_t get_max_armor();
+	void set_max_armor(double value);
+	double get_max_armor();
 
 	double get_armor_pct();
 	double get_tank_pct();
 
-	void repair(int64_t amount);
-	void damage(int64_t amount);
+	void repair(double amount);
+	void damage(double amount);
 
 	void set_gen_hit_delay(double value);
 	double get_gen_hit_delay();
